Transactions ft/deposit

+++ src/modules/product/product.controller.js
@@ -1,5 +1,5 @@
 import Product from './product.model.js';
-import { validateUserRequest } from "../../helpers/controller-checks.js"
+import { validateUserRequest, validateAdminRequest } from "../../helpers/controller-checks.js"
 import { handleResponse, handleResponseWithMessage } from "../../helpers/handle-resp.js"
 import { logger } from "../../helpers/logger.js";
 
@@ -7,20 +7,20 @@ export const createProductServiceDiscount = async (req, res) => {
     logger.info('Starting product/Service creation');
     const { type, name, description, enterprice, discount } = req.body;
 
-    await validateUserRequest(req, res);
+    await validateAdminRequest(req, res);
     handleResponse(res, Product.create({ type, name, description, enterprice, discount }));
 }
 
 export const getAllProductsAndServices = async (req, res) => {
     logger.info('Getting all products/Services');
-    await validateUserRequest(req, res);
+    await validateAdminRequest(req, res);
     handleResponse(res, Product.find({}));
 }
 
 export const deleteProductService = async (req, res) => {
     logger.info('Starting product/Service deletion');
     const { id } = req.body;
-    await validateUserRequest(req, res);
+    await validateAdminRequest(req, res);
     handleResponse(res, Product.findByIdAndDelete(id));
 }
